move `extractNodeRequestData` (and associated tests) from utils to node package, rename
