+++ packages/node/src/handlers.ts
@@ -2,18 +2,13 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { captureException, getCurrentHub, startTransaction, withScope } from '@sentry/core';
 import { extractTraceparentData, Span } from '@sentry/tracing';
-import { Event, Transaction } from '@sentry/types';
-import {
-  extractNodeRequestData,
-  forget,
-  isPlainObject,
-  isString,
-  logger,
-  stripUrlQueryAndFragment,
-} from '@sentry/utils';
+import { Event, ExtractedNodeRequestData, Transaction } from '@sentry/types';
+import { forget, isPlainObject, isString, logger, normalize, stripUrlQueryAndFragment } from '@sentry/utils';
+import * as cookie from 'cookie';
 import * as domain from 'domain';
 import * as http from 'http';
 import * as os from 'os';
+import * as url from 'url';
 
 import { NodeClient } from './client';
 import { flush } from './sdk';
@@ -179,6 +174,97 @@ function extractUserData(
   return extractedUser;
 }
 
+/** Default request keys that'll be used to extract data from the request */
+const DEFAULT_REQUEST_KEYS = ['cookies', 'data', 'headers', 'method', 'query_string', 'url'];
+
+/**
+ * Normalizes data from the request object, accounting for framework differences.
+ *
+ * @param req The request object from which to extract data
+ * @param keys An optional array of keys to include in the normalized data. Defaults to DEFAULT_REQUEST_KEYS if not
+ * provided.
+ * @returns An object containing normalized request data
+ */
+export function extractRequestData(
+  req: { [key: string]: any },
+  keys: string[] = DEFAULT_REQUEST_KEYS,
+): ExtractedNodeRequestData {
+  const requestData: { [key: string]: any } = {};
+
+  // headers:
+  //   node, express: req.headers
+  //   koa: req.header
+  const headers = (req.headers || req.header || {}) as {
+    host?: string;
+    cookie?: string;
+  };
+  // method:
+  //   node, express, koa: req.method
+  const method = req.method;
+  // host:
+  //   express: req.hostname in > 4 and req.host in < 4
+  //   koa: req.host
+  //   node: req.headers.host
+  const host = req.hostname || req.host || headers.host || '<no host>';
+  // protocol:
+  //   node: <n/a>
+  //   express, koa: req.protocol
+  const protocol =
+    req.protocol === 'https' || req.secure || ((req.socket || {}) as { encrypted?: boolean }).encrypted
+      ? 'https'
+      : 'http';
+  // url (including path and query string):
+  //   node, express: req.originalUrl
+  //   koa: req.url
+  const originalUrl = (req.originalUrl || req.url || '') as string;
+  // absolute url
+  const absoluteUrl = `${protocol}://${host}${originalUrl}`;
+
+  keys.forEach(key => {
+    switch (key) {
+      case 'headers':
+        requestData.headers = headers;
+        break;
+      case 'method':
+        requestData.method = method;
+        break;
+      case 'url':
+        requestData.url = absoluteUrl;
+        break;
+      case 'cookies':
+        // cookies:
+        //   node, express, koa: req.headers.cookie
+        //   vercel, sails.js, express (w/ cookie middleware): req.cookies
+        // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+        requestData.cookies = req.cookies || cookie.parse(headers.cookie || '');
+        break;
+      case 'query_string':
+        // query string:
+        //   node: req.url (raw)
+        //   express, koa: req.query
+        // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+        requestData.query_string = url.parse(originalUrl || '', false).query;
+        break;
+      case 'data':
+        if (method === 'GET' || method === 'HEAD') {
+          break;
+        }
+        // body data:
+        //   node, express, koa: req.body
+        if (req.body !== undefined) {
+          requestData.data = isString(req.body) ? req.body : JSON.stringify(normalize(req.body));
+        }
+        break;
+      default:
+        if ({}.hasOwnProperty.call(req, key)) {
+          requestData[key] = (req as { [key: string]: any })[key];
+        }
+    }
+  });
+
+  return requestData;
+}
+
 /**
  * Options deciding what parts of the request to use when enhancing an event
  */
@@ -222,10 +308,10 @@ export function parseRequest(event: Event, req: ExpressRequest, options?: ParseR
   }
 
   if (options.request) {
-    // if the option value is `true`, use the default set of keys by not passing anything to `extractNodeRequestData()`
+    // if the option value is `true`, use the default set of keys by not passing anything to `extractRequestData()`
     const extractedRequestData = Array.isArray(options.request)
-      ? extractNodeRequestData(req, options.request)
-      : extractNodeRequestData(req);
+      ? extractRequestData(req, options.request)
+      : extractRequestData(req);
     event.request = {
       ...event.request,
       ...extractedRequestData,
