+++ packages/node/test/handlers.test.ts
@@ -8,7 +8,7 @@ import * as net from 'net';
 
 import { Event, Request, User } from '../src';
 import { NodeClient } from '../src/client';
-import { ExpressRequest, parseRequest, tracingHandler } from '../src/handlers';
+import { ExpressRequest, extractRequestData, parseRequest, tracingHandler } from '../src/handlers';
 
 describe('parseRequest', () => {
   let mockReq: { [key: string]: any };
@@ -345,4 +345,179 @@ describe('tracingHandler', () => {
       done();
     });
   });
-}); // end describe('tracingHandler')
+});
+
+describe('extractRequestData()', () => {
+  describe('default behaviour', () => {
+    test('node', () => {
+      expect(
+        extractRequestData({
+          headers: { host: 'example.com' },
+          method: 'GET',
+          secure: true,
+          originalUrl: '/',
+        }),
+      ).toEqual({
+        cookies: {},
+        headers: {
+          host: 'example.com',
+        },
+        method: 'GET',
+        query_string: null,
+        url: 'https://example.com/',
+      });
+    });
+
+    test('degrades gracefully without request data', () => {
+      expect(extractRequestData({})).toEqual({
+        cookies: {},
+        headers: {},
+        method: undefined,
+        query_string: null,
+        url: 'http://<no host>',
+      });
+    });
+  });
+
+  describe('cookies', () => {
+    it('uses `req.cookies` if available', () => {
+      expect(
+        extractRequestData(
+          {
+            cookies: { foo: 'bar' },
+          },
+          ['cookies'],
+        ),
+      ).toEqual({
+        cookies: { foo: 'bar' },
+      });
+    });
+
+    it('parses the cookie header', () => {
+      expect(
+        extractRequestData(
+          {
+            headers: {
+              cookie: 'foo=bar;',
+            },
+          },
+          ['cookies'],
+        ),
+      ).toEqual({
+        cookies: { foo: 'bar' },
+      });
+    });
+
+    it('falls back if no cookies are defined', () => {
+      expect(extractRequestData({}, ['cookies'])).toEqual({
+        cookies: {},
+      });
+    });
+  });
+
+  describe('data', () => {
+    it('includes data from `req.body` if available', () => {
+      expect(
+        extractRequestData(
+          {
+            method: 'POST',
+            headers: { 'Content-Type': 'application/x-www-form-urlencoded' },
+            body: 'foo=bar',
+          },
+          ['data'],
+        ),
+      ).toEqual({
+        data: 'foo=bar',
+      });
+    });
+
+    it('encodes JSON body contents back to a string', () => {
+      expect(
+        extractRequestData(
+          {
+            method: 'POST',
+            headers: { 'Content-Type': 'application/json' },
+            body: { foo: 'bar' },
+          },
+          ['data'],
+        ),
+      ).toEqual({
+        data: '{"foo":"bar"}',
+      });
+    });
+  });
+
+  describe('query_string', () => {
+    it('parses the query parms from the url', () => {
+      expect(
+        extractRequestData(
+          {
+            headers: { host: 'example.com' },
+            secure: true,
+            originalUrl: '/?foo=bar',
+          },
+          ['query_string'],
+        ),
+      ).toEqual({
+        query_string: 'foo=bar',
+      });
+    });
+
+    it('gracefully degrades if url cannot be determined', () => {
+      expect(extractRequestData({}, ['query_string'])).toEqual({
+        query_string: null,
+      });
+    });
+  });
+
+  describe('url', () => {
+    test('express/koa', () => {
+      expect(
+        extractRequestData(
+          {
+            host: 'example.com',
+            protocol: 'https',
+            url: '/',
+          },
+          ['url'],
+        ),
+      ).toEqual({
+        url: 'https://example.com/',
+      });
+    });
+
+    test('node', () => {
+      expect(
+        extractRequestData(
+          {
+            headers: { host: 'example.com' },
+            secure: true,
+            originalUrl: '/',
+          },
+          ['url'],
+        ),
+      ).toEqual({
+        url: 'https://example.com/',
+      });
+    });
+  });
+
+  describe('custom key', () => {
+    it('includes the custom key if present', () => {
+      expect(
+        extractRequestData(
+          {
+            httpVersion: '1.1',
+          },
+          ['httpVersion'],
+        ),
+      ).toEqual({
+        httpVersion: '1.1',
+      });
+    });
+
+    it('gracefully degrades if the custom key is missing', () => {
+      expect(extractRequestData({}, ['httpVersion'])).toEqual({});
+    });
+  });
+});
